Switch to Rust TLS, remove attohttpc, fix clap error
This commit replaces the `attohttpc` dependency with the `blocking` feature on `reqwest`.

Additionally, `reqwest` is compiled without default features, and with the `rustls-tls` feature enabled.
(For some reason, `openssl-sys` wouldn't recognize my OpenSSL installation, so this should
improve compatibility, especially for Windows users.)

Lastly, the executable failed to start due to a duplicate short argument (-t in `cargo skyline new`), so
the `--template-git-branch` short version was changed to `-b`.
I couldn't reproduce this on the latest release, so it is possible that it might have been introduced by a
`clap` update.